refactor: #15 AlarmSettingViewController를 View 파일로 분리
