+++ Alarmy-iOS/Alarmy-iOS/App/SceneDelegate.swift
@@ -14,7 +14,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         
-        let viewController = AlarmyTabBarController()
+        let viewController = AlarmSettingViewController()
         let navigationController = UINavigationController(rootViewController: viewController)
         
         let window = UIWindow(windowScene: windowScene)
